+++ src/api.rs
@@ -235,6 +235,7 @@ impl Ludusavi {
                 &self.steam_shortcuts,
                 self.config.backup.only_constructive,
                 self.config.scan.emulator_saves,
+                &self.config.scan.emulator_save_templates,
             );
             let ignored = !&self.config.is_game_enabled_for_backup(name) && !games_specified && !include_disabled;
             let decision = if ignored {